apps/worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
psycopg2-binary==2.9.9
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==10.4.0
httpx==0.27.2
tiktoken==0.8.0
langextract[openai]
google-cloud-storage>=2.14.0
google-cloud-documentai>=2.20.0
langdetect>=1.0.9
xlrd==2.0.1
